+++ CHANGELOG.md
@@ -14,6 +14,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Fixed** for any bug fixes.
 - **Security** in case of vulnerabilities.
 
+## [0.3.0] - 2024-01-14
+
+### Added
+
+- AuthError to KnownError
+
+### Changed
+
+- Login now uses username and password
+- Better logic for handling errors in generalActions
+
 ## [0.2.0] - 2024-01-13
 
 ### Added